Add parsing of ahead and behid
We have to split to the code if we want some function used
in test.c.

Meson default test target uses a a technique to perturbate
malloc:

$ ./builddir/mytest
hello test
Branch: (null), Ahead: 5, Behind: 20

$ MALLOC_PERTURB_=126 ./builddir/mytest
hello test
Segmentation fault

Hence, we also had to add a call to parsing a branch name, so
it will not be null.